move humanizeText method to shared class
