+++ src/Web/opencatapultweb/src/app/core/services/text-helper.service.ts
@@ -1,0 +1,17 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class TextHelperService {
+
+  constructor() { }
+
+  humanizeText(text: string): string {
+    return text
+        .replace(/([A-Z])/g, ' $1')
+        .replace(/^[\s_]+|[\s_]+$/g, '')
+        .replace(/[_\s]+/g, ' ')
+        .replace(/^[\s-]+|[\s-]+$/g, '')
+        .replace(/[-\s]+/g, ' ')
+        .replace(/^[a-z]/, function(m) { return m.toUpperCase(); });
+  }
+}
