+++ src/Web/opencatapultweb/src/app/project/components/project-info-form/project-info-form.component.ts
@@ -3,6 +3,7 @@ import { FormBuilder, Validators, FormGroup } from '@angular/forms';
 import { ProjectDto } from '@app/core';
 import { Subject } from 'rxjs';
 import { debounceTime, distinctUntilChanged, map } from 'rxjs/operators';
+import { TextHelperService } from '@app/core/services/text-helper.service';
 
 @Component({
   selector: 'app-project-info-form',
@@ -18,7 +19,8 @@ export class ProjectInfoFormComponent implements OnInit, OnChanges {
   private projectName = new Subject<string>();
 
   constructor(
-    private fb: FormBuilder
+    private fb: FormBuilder,
+    private textHelperService: TextHelperService
     ) { }
 
   ngOnInit() {
@@ -72,7 +74,7 @@ export class ProjectInfoFormComponent implements OnInit, OnChanges {
 
       this.projectInfoForm.patchValue({
         name: projectName,
-        displayName: this.humanize(projectName)
+        displayName: this.textHelperService.humanizeText(projectName)
       });
     });
   }
@@ -80,15 +82,5 @@ export class ProjectInfoFormComponent implements OnInit, OnChanges {
   isFieldInvalid(field: string, errorCode: string) {
     const control = this.projectInfoForm.get(field);
     return control.invalid && control.errors && control.getError(errorCode);
-  }
-
-  private humanize(str) {
-    return str
-        .replace(/([A-Z])/g, ' $1')
-        .replace(/^[\s_]+|[\s_]+$/g, '')
-        .replace(/[_\s]+/g, ' ')
-        .replace(/^[\s-]+|[\s-]+$/g, '')
-        .replace(/[-\s]+/g, ' ')
-        .replace(/^[a-z]/, function(m) { return m.toUpperCase(); });
   }
 }
