+++ src/Web/opencatapultweb/src/app/project/project/project.component.css
@@ -13,7 +13,7 @@ mat-expansion-panel-header {
 
 .is-active {
   background-color: #f5f5f5;
-  border-left: solid 2px #3f51b5;
+  border-left: solid 2px #faa438;
 }
 
 .button {
